Change bar chart segment size to a float
As the height of a segment may not perfectly align to an int, using an
int means a loss of precision in the scale of the bars. This becomes
apparent as the maximum number of segments needed grows. Eventually,
this int would hit zero, and no bars would be drawn. Using a float
allows the bars to utilize all available space and avoid hitting a zero
segment height (up to a certain point. I have yet to determine when the
precision of a float is too small that the segment height would become
zero.).